+++ packages/oidc-verifier-jose/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @orpc-ws/oidc-verifier-jose
 
+## 0.13.0
+
+### Patch Changes
+
+- @orpc-ws/server@0.13.0
+
 ## 0.12.3
 
 ### Patch Changes