+++ packages/orpc-ws-shared/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @orpc-ws/shared
 
+## 0.13.0
+
 ## 0.12.3
 
 ### Patch Changes